+++ src/main/java/org/tlgcohort/manvswild/Commands/CommandEngine.java
@@ -1,6 +1,6 @@
 package org.tlgcohort.manvswild.Commands;
-import org.tlgcohort.manvswild.Thing.LocationPOJO;
-import org.tlgcohort.manvswild.Thing.Player;
+import org.tlgcohort.manvswild.Things.LocationPOJO;
+import org.tlgcohort.manvswild.Things.Player;
 import java.util.List;
 import java.util.Scanner;
