+++ src/main/java/org/tlgcohort/manvswild/GridOne/GridOneEngine.java
@@ -1,24 +1,27 @@
-package org.tlgcohort.manvswild.GridTwo;
+package org.tlgcohort.manvswild.GridOne;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import org.tlgcohort.manvswild.Thing.Map;
-import org.tlgcohort.manvswild.Thing.Location;
+import org.tlgcohort.manvswild.Things.Map;
+import org.tlgcohort.manvswild.Things.Location;
 import org.tlgcohort.manvswild.DataManagement.DataManagement;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
-public class GridTwoEngine {
-    public GridTwo gridTwoEngine() throws IOException { //returning a gridTwo object to be able to access GridTwo
+public class GridOneEngine {
+
+    private String name;
+
+    public GridOne gridOneEngine() throws IOException {
         DataManagement dm = new DataManagement();
-        File file = new File("src/main/resources/gridTwo.json");    //grabbing appropriate file
+        File file = new File("src/main/resources/gridOne.json");
         JsonNode node = dm.parse(file);
-        GridTwo g2 = dm.fromJson(node,GridTwo.class);
-        return g2;
+        GridOne g1 = dm.fromJson(node,GridOne.class);
+        return g1;
     }
 
-    public List<Location> gridTwoInformation() throws IOException {
+    public List<Location> gridOneInformation() throws IOException {
         DataManagement dm = new DataManagement();
         File file = new File("src/main/resources/locations.json");
         JsonNode node = dm.parse(file);
@@ -26,4 +29,5 @@ public class GridTwoEngine {
         List<Location> output = gameMap.getLocations();
         return output;
     }
+
 }
