+++ src/main/java/org/tlgcohort/manvswild/Things/LocationPOJO.java
@@ -1,4 +1,4 @@
-package org.tlgcohort.manvswild.Thing;
+package org.tlgcohort.manvswild.Things;
 
 public class LocationPOJO {
