+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,1 @@
+Main-Class: org.tlgcohort.manvswild.MainDriver
